requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic-settings==2.12.0
SQLAlchemy==2.0.23
alembic==1.12.1
psycopg[binary]==3.3.2
PyJWT==2.8.0
httpx==0.25.2
